Sexting choice dialogue width, 'getting close'  dialogue
Adds nsfw_screens to provide an override for the screen choice(item) function, itself defined in the MAS files. This override means that if you put ` (sextchoice)` (space included into a dialogue option, it will remove that and apply a new style to the buttons, which have double the usual width.
This has a two-fold improvement; for one, lengthly sexting options will no longer go off the top of the screen. Secondly, the added space means there will be room for more choices down the line (e.g. a finish early button).

This also includes a few quick variations added to monika's "getting close" text in nsfw_sexting_finale. In future, this can be moved into nsfw_main to avoid clutter, but it exists here as a proof of concept.