clothes check and successful ending
Fixed the lingerie check on start and the changing conditions so she will not change if she was already wearing lingerie or (in case of sexy start) being nude on start.

Also introduced a "success" condition so that Monika can now end a successful sexting session with "I love you" and the player can reply.