Minor bugfixes and new cooldown setting
Fixed a minor bug in mas_nsfw.rpy and changed the conditions for sexting teaser and attempt so a new cooldown of 48 hours could be added. When Monika tries to restart an interrupted session, the reattempt cooldown will now always increase by 5 minutes after each failed attempt regardless of the selected value. Also if her first attempt to start a session was denied, she will now always wait one third of the base cooldown for the next attempt until you give in or she gives up.